refactor: recoil의 atom, selector 분리
